+++ src/Places/CustomFile.php
@@ -7,8 +7,7 @@
  * - 'custom_file_path' => '/your/absolute/path/to/creds.php',
  *
  * Hint:
- * - File must be a PHP file.
- * - This file must be included in WordPress to get loaded.
+ * - This file will be embedded by this package on load.
  *
  * @package crypt-for-wordpress
  */
@@ -23,12 +22,12 @@ use CryptForWordPress\Helper;
 use CryptForWordPress\Place_Base;
 
 /**
- * Object to handle the wp-config.php as place to save the key.
+ * Object to handle a custom file as place to save the key.
  */
 class CustomFile extends Place_Base {
 
 	/**
-	 * Name of the method.
+	 * Name of the place.
 	 *
 	 * @var string
 	 */
@@ -54,6 +53,11 @@ class CustomFile extends Place_Base {
 
 		// bail if no file path is given.
 		if ( empty( $config['custom_file_path'] ) ) {
+			return false;
+		}
+
+		// bail if given value is not a string.
+		if ( ! is_string( $config['custom_file_path'] ) ) {
 			return false;
 		}
 
@@ -93,5 +97,36 @@ class CustomFile extends Place_Base {
 
 		// save the changed wp-config.php.
 		$wp_filesystem->put_contents( $config['custom_file_path'], $wp_config_php_content ); // @phpstan-ignore argument.type
+	}
+
+	/**
+	 * Load this places environments before the crypt method is used.
+	 *
+	 * @return void
+	 */
+	public function load(): void {
+		// get the configuration.
+		$config = $this->get_crypt_obj()->get_config();
+
+		// bail if no file path is given.
+		if ( empty( $config['custom_file_path'] ) ) {
+			return;
+		}
+
+		// bail if given value is not a string.
+		if ( ! is_string( $config['custom_file_path'] ) ) {
+			return;
+		}
+
+		// get the WP_Filesystem object.
+		$wp_filesystem = Helper::get_wp_filesystem();
+
+		// bail if the file does not exist.
+		if ( ! $wp_filesystem->exists( $config['custom_file_path'] ) ) {
+			return;
+		}
+
+		// embed the given file.
+		require_once $config['custom_file_path'];
 	}
 }