+++ src/Places/WpConfig.php
@@ -20,7 +20,7 @@ use CryptForWordPress\Place_Base;
 class WpConfig extends Place_Base {
 
 	/**
-	 * Name of the method.
+	 * Name of the place.
 	 *
 	 * @var string
 	 */
@@ -112,7 +112,7 @@ class WpConfig extends Place_Base {
 	}
 
 	/**
-	 * Uninstall this method.
+	 * Uninstall this place.
 	 *
 	 * @param string $constant The constant to use during the uninstallation.
 	 * @return void