+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added 2 new places for the key
+-> an environment variable from .env-file
+-> a server variable in the hosting
+- Added the function debug() to return the actual configuration for debugging purposes
+- Added some documentation
+
+### Changed
+
+- Load a given custom file before the method will be loaded to embed its constants
+
 ## [1.2.1] - 07.04.2026
 
 ### Fixed